2002-07-12  Marc Mulcahy <marc.mulcahy@sun.com>
	* src/fileinfo.c: Fixed rounding errors in finding audio out queue points.

2002-07-12  Marc Mulcahy <marc@plbb.net>

	* src/mixerchannel.c src/mixerchannel.h src/vorbismixerchannel.c
	src/mixer.c src/mixer.h: Added mixer channel level and fading support.
	Added an end of data flag so we can appropriately dispose of dead
	channels.


git-svn-id: svn://svn.code.sf.net/p/prsradio/code/trunk@12 410355f2-c4ea-45de-a5b5-301721f902b9

+++ prs/src/fileinfo.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 #include <vorbis/vorbisfile.h>
 #include "fileinfo.h"
@@ -103,7 +104,7 @@ get_vorbis_audio_out (OggVorbis_File *vf,
 	break;
       ptr--;
       }
-  audio_out = seek_time + (ptr-(short *)buffer)/vi->rate/vi->channels;
+  audio_out = seek_time + (double) (ptr-(short *)buffer)/vi->rate/vi->channels;
   free (buffer);
   return audio_out;
 }
